backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg[binary]==3.2.10
pydantic==2.10.6
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-dotenv==1.0.1
